requirements.txt
cryptography
pyperclip
Pillow
reportlab
networkx
matplotlib
scipy
qrcode[pil]
pyzbar
requests
ddgs
squarify
psutil
python-whois
python-nmap
pywifi
comtypes
scapy
pandas
seaborn
PyYAML
GitPython
numpy